Move ReportFilteredTransactions to its own file
