+++ cmd/filtering-report/api/api.go
@@ -4,20 +4,13 @@
 package api
 
 import (
-	"context"
-	"encoding/json"
-	"errors"
 	"net/http"
 
-	"github.com/aws/aws-sdk-go-v2/service/sqs"
-
-	"github.com/ethereum/go-ethereum/log"
 	"github.com/ethereum/go-ethereum/node"
 	"github.com/ethereum/go-ethereum/p2p"
 	"github.com/ethereum/go-ethereum/rpc"
 
 	"github.com/offchainlabs/nitro/execution/gethexec"
-	"github.com/offchainlabs/nitro/execution/gethexec/addressfilter"
 	"github.com/offchainlabs/nitro/util/sqsclient"
 )
 
@@ -29,30 +22,6 @@ func NewFilteringReportAPI(sqsClient *sqsclient.QueueClient) *FilteringReportAPI
 	return &FilteringReportAPI{
 		sqsClient: sqsClient,
 	}
-}
-
-func (a *FilteringReportAPI) ReportFilteredTransactions(ctx context.Context, reports []addressfilter.FilteredTxReport) error {
-	if a.sqsClient == nil {
-		return errors.New("SQS client not configured")
-	}
-	log.Debug("Sending filtered transaction reports to SQS", "count", len(reports))
-	for _, report := range reports {
-		body, err := json.Marshal(report)
-		if err != nil {
-			return err
-		}
-		bodyStr := string(body)
-		_, err = a.sqsClient.SendMessage(ctx, &sqs.SendMessageInput{
-			QueueUrl:    &a.sqsClient.QueueURL,
-			MessageBody: &bodyStr,
-		})
-		if err != nil {
-			log.Error("Failed to send filtered transaction report to SQS", "txHash", report.TxHash.Hex(), "err", err)
-			return err
-		}
-		log.Debug("Successfully sent filtered transaction report to SQS", "txHash", report.TxHash.Hex())
-	}
-	return nil
 }
 
 var DefaultStackConfig = node.Config{
